Refactor styles and clean up code
