+++ src/components/admin/AdminMenu.styles.tsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const AdminMenuContainer = styled.div`
+  background-color: rgba(255, 255, 255, 0.6);
+  display: flex;
+  justify-content:center;
+  text-transform: uppercase;
+
+  a {
+    color: green;
+    padding: 10px;
+    text-decoration: none;
+  }
+  a:hover {
+    color: white;
+  }
+`;
