+++ src/components/admin/AdminMenu.tsx
@@ -1,7 +1,11 @@
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
 
-export default function AdminMenu() {
+//Styles
+import { AdminMenuContainer } from './AdminMenu.styles';
+
+
+
+const AdminMenu = () => {
   return(
     <AdminMenuContainer>
       <Link to="/admin/players">Players</Link>
@@ -10,18 +14,4 @@ export default function AdminMenu() {
   )
 }
 
-const AdminMenuContainer = styled.div`
-  background-color: rgba(255, 255, 255, 0.6);
-  display: flex;
-  justify-content:center;
-  text-transform: uppercase;
-
-  a {
-    color: green;
-    padding: 10px;
-    text-decoration: none;
-  }
-  a:hover {
-    color: white;
-  }
-`;
+export default AdminMenu;
