+++ src/components/admin/Filters.styles.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const FilterContainer = styled.div`
+  display: flex;
+  background-color: white;
+  flex-wrap: wrap;
+  justify-content: center;
+  width: 100%;
+  > div {
+    margin: 5px;
+  }
+`;
