+++ src/components/admin/Player.tsx
@@ -1,15 +1,19 @@
 import axios from 'axios';
 import { useParams, useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
 
+//Components
 import Button from "@mui/material/Button";
 import PlayerWeekList from './PlayerWeekList';
 
+//Global functions
 import { fetchAdminUsers } from '../../functions';
 
+//Styles
+import { PlayerInfoContainer, PlayerInfoSection } from './Player.styles';
 
 
-export default function Player(props: any) {
+
+const Player = (props: any) => {
   const { users, token, setUsers } = props;
   const { username } = useParams();
   const navigate = useNavigate();
@@ -45,44 +49,4 @@ export default function Player(props: any) {
   )
 }
 
-const PlayerInfoSection = styled.div`
-  width: 100%;
-  padding: 20px;
-  background-color: lightcyan;
-  color:rgb(6, 128, 55);
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  max-height: 300px;
-  position: relative;
-
-  .tag {
-    margin-top: 0;
-    padding: 5px 20px;
-    border-radius: 5px;
-    color: white;
-  }
-
-  .active-tag {
-    background-color: rgb(6, 128, 55);
-  }
-
-  .inactive-tag {
-    background-color: grey;
-  }
-
-  ul {
-    list-style-type: none;
-    text-align: left;
-  }
-`
-
-const PlayerInfoContainer = styled.div`
-  margin-top: 10px;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-`
+export default Player;
