+++ src/components/admin/PlayerList.tsx
@@ -1,12 +1,15 @@
 import { useState } from 'react';
 
-import styled from 'styled-components';
-
+//Components
 import PlayerListRow from './PlayerListRow';
 import PlayerListHeader from './PlayerListHeader';
 
+//Styles
+import { PlayersListContainer } from './PlayerList.styles';
 
-export default function PlayerList(props: any) {
+
+
+const PlayerList = (props: any) => {
   const [isCheckAll, setIsCheckAll] = useState(false);
   const [isCheck, setIsCheck] = useState<{ isChecked: string }[]>([]);
 
@@ -60,29 +63,4 @@ export default function PlayerList(props: any) {
   )
 }
 
-const PlayersListContainer = styled.table`
-  margin: 0 auto;
-  background-color: white;
-  width: 100%;
-  padding: 5px;
-
-  th {
-    border-bottom: 1px solid;
-    text-align: center;
-  }
-  td {
-    height: 20px;
-  }
-  tbody tr:nth-child(odd) {
-    background-color: lightgrey;
-
-    :hover {
-      background-color: green;
-      color: white;
-    }
-  }
-  tbody tr:hover {
-    color: white;
-    background-color: green;
-  }
-`
+export default PlayerList;
