+++ src/components/admin/PlayerListColumnHeader.styles.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const TableHeaderStyle = styled.p`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  padding: 0;
+  margin: 0;
+`;
