+++ src/components/admin/PlayerListRow.styles.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const TableData = styled.td`
+  .link {
+    cursor: pointer;
+  }
+
+  .inactive {
+    color: black;
+    font-weight: 100;
+    font-style: italic;
+  }
+`;
