+++ src/components/admin/functions/index.tsx
@@ -1,3 +1,5 @@
+import axios from 'axios';
+
 import { teams } from '../../../data/teams';
 
 export const filterUsersByUsername = (users: any, value: any) => {
@@ -124,4 +126,20 @@ export const getTeamsArray = () => {
     teamNames.push(team.name)
   })
   return teamNames;
+}
+
+export async function eliminateUsers(token: any, users: any) {
+  try {
+    const response = await axios.patch(`${process.env.REACT_APP_SERVER}/admin/eliminate`, {
+      username: users
+    },{
+      headers: {
+        'Authorization': `Bearer ${token}`
+      }
+    });
+    const data = await response.data;
+    return data;
+  } catch (error) {
+    console.log(error);
+  }
 }
